+++ changelog.md
@@ -48,6 +48,11 @@
     * Added Kotlin wrappers for `.fromOther` and `.toOther` that allow you to pass a class type rather than a creator. e.g. `.toOther<Fuse>()` instead
       of `.toOther(::Fuse)` or `.toOther( { Fuse(it) } )`.
 * Added `+` and `-` operators to `PhaseCode` and `SinglePhaseKind`.
+* `TraversalQueue` now has `addAll` methods taking either a collection or varargs, which by default will just call `add` for each item, but can be overridden if
+  there is an `addAll` available on the underlying queue implementation.
+* `Traversal` has two new helper methods:
+    * `ifNotStopping`: Adds a step action that is only called if the traversal is not stopping on the item.
+    * `ifStopping`: Adds a step action that is only called if the traversal is stopping on the item.
 
 ### Fixes